+++ lib/src/midi/device/worlde_easycontrol9.rs
@@ -12,9 +12,9 @@ pub enum Input {
     MainButton(u8, bool),
     CtrlButton(u8, bool),
     BankButton(u8, bool),
-    Slider(u8, f64),
-    Knob(u8, f64),
-    Fader(f64),
+    Slider(u8, f32),
+    Knob(u8, f32),
+    Fader(f32),
     Encoder(bool),
     Bank(u8),
 }
@@ -29,11 +29,11 @@ impl MidiDevice for WorldeEasyControl9 {
                 let cc = raw[1];
                 let state = raw[2];
                 let on = state == 127;
-                let fl = state as f64 / 126.0;
+                let fl = state as f32 / 126.0;
 
                 match cc {
                     1..=2 => Input::BankButton(cc - 1, on),
-                    9 => Input::Fader((std::cmp::max(state, 1) - 1) as f64 / 126.0),
+                    9 => Input::Fader((std::cmp::max(state, 1) - 1) as f32 / 126.0),
                     14..=22 => Input::Knob(cc - 14, fl),
                     23..=31 => Input::MainButton(cc - 23, on),
                     32..=40 => Input::Slider(cc - 32, fl),
