+++ lib/src/num/mod.rs
@@ -11,8 +11,8 @@ pub use interp::Interp;
 pub use range::Range;
 
 /// τ, a full circle.
-pub const TAU: f64 = std::f64::consts::TAU;
+pub const TAU: f32 = std::f32::consts::TAU;
 /// τ/2, a half-circle.
-pub const TAU_2: f64 = TAU / 2.0;
+pub const TAU_2: f32 = TAU / 2.0;
 /// τ/4, a quarter-circle.
-pub const TAU_4: f64 = TAU / 4.0;
+pub const TAU_4: f32 = TAU / 4.0;
